BCEfficiencyFitter: - rewritten calculation of data points and their error-bars - changed default to mode + smallest 68% interval - added posibility to also plot mean + rms and median + central 68% interval - changed some warnings to errors - code cleanup BCH1D: - small update in GetSmallestInterval()

+++ src/BCMath.cxx
@@ -110,7 +110,7 @@ double LogBinomFactor(int n, int k)
       return log((double) n);
 
    // if no approximation needed
-   if ( n < (int) BCMATH_NFACT_ALIMIT || (n < (int) BCMath::nCacheFact &&  (n - k) < 10) )
+   if ( n < BCMATH_NFACT_ALIMIT || (n < (int) BCMath::nCacheFact &&  (n - k) < 10) )
       return LogNoverK(n, k);
 
    // calculate final log(n over k) using approximations if necessary
